add Node.js v20 support (#28)
* add Node.js v20 support

Since Node.js v20 moves loaders to a separate thread, we can no longer
depend on loading the modules in the loader to get the exports. This is
needed to add our mutable proxy. For Node.js 20, exports are now
retrieved via parsing.

To reduce startup overhead on older versions of Node.js, the previous
method of getting exports is used, avoiding loading and using parsers.

* make stacked loaders work

* restrict get-esm-exports test to node 20

* remove 'unsupported'

* normalize file URLs to paths (important on windows)

* remove trailing \r in esm exports test