feat: Add `Hook` named export (#88)
`import-in-the-middle` has some ESM interop issues.

From
[here](https://github.com/open-telemetry/opentelemetry-js/pull/4546#issue-2186767832):

> Specifically, namespace imports are required by the ESM spec to be
objects. IITM's top-level export is not an object, but a callable
function

This causes
[issues](https://github.com/open-telemetry/opentelemetry-js/issues/4717)
with some bundlers. Obviously bundled imports will not go through
`import-in-the-middle` but using it should not inhibit bundlers as iitm
can still hook build-in modules.

This adds a `Hook` named export and changes one cjs and esm test to
use that new export.