fix: use createRequire to load hook.js
If import is used to load hook.js, then any loaders that assign
`result.source` (as indicated by the Node.js docs) on CommonJS files
will cause a not-implemented error to be thrown inside Node.js. This is
probably a bug in Node.js, but this ought to be an appropriate
workaround (espectially since IITM supports old versions of Node.js).