fix: properly hook submodule package exports
This requires passing the specifier down the chain to the hook function,
in order to apply the hook function when it matches.

By the time the hook is being matched, the submodule specifier is
already overwritten to a full `file://` URL, and appears as if the user
is requesting an arbitrary internal within the package by absolute path.
Since the module detected in that path does not match any of the module
specifiers, the hook would previously be skipped unless the `internals:
true` option was set, _and_ the submodule export path matched the file
subpath exactly.

With the original specifier present, the hook can be applied, by
matching the specifier requested by the user to the submodule specifier
provided at hook creation.

This is far less brittle as a hook author than using `internals: true`,
because it does not depend on the internal structure of the package, but
rather only the public API surface, which is much less likely to change.